Continuation of 64-bit GUID Migration

+++ src/server/game/AuctionHouseBot/AuctionHouseBot.cpp
@@ -82,7 +82,7 @@ bool AuctionBotConfig::Initialize()
                 do
                 {
                     Field* fields = result->Fetch();
-                    _AHBotCharacters.push_back(fields[0].GetUInt32());
+                    _AHBotCharacters.push_back(fields[0].GetUInt64());
                     ++count;
                 } while (result->NextRow());
             }
