+++ src/server/game/Tools/PlayerDump.cpp
@@ -614,15 +614,15 @@ void PlayerDumpWriter::PopulateGuids(ObjectGuid::LowType guid)
             switch (baseTable.StoredType)
             {
                 case GUID_TYPE_ITEM:
-                    if (ObjectGuid::LowType itemLowGuid = (*result)[0].GetUInt32())
+                    if (ObjectGuid::LowType itemLowGuid = (*result)[0].GetUInt64())
                         _items.insert(itemLowGuid);
                     break;
                 case GUID_TYPE_MAIL:
-                    if (ObjectGuid::LowType mailLowGuid = (*result)[0].GetUInt32())
+                    if (ObjectGuid::LowType mailLowGuid = (*result)[0].GetUInt64())
                         _mails.insert(mailLowGuid);
                     break;
                 case GUID_TYPE_PET:
-                    if (ObjectGuid::LowType petLowGuid = (*result)[0].GetUInt32())
+                    if (ObjectGuid::LowType petLowGuid = (*result)[0].GetUInt64())
                         _pets.insert(petLowGuid);
                     break;
                 case GUID_TYPE_EQUIPMENT_SET:
